add Spring Boot

+++ src/main/java/ru/clevertec/service/CardServiceImpl.java
@@ -1,54 +1,55 @@
 package ru.clevertec.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.clevertec.dto.CardDto;
 import ru.clevertec.entity.Card;
-import ru.clevertec.exception.RepositoryException;
 import ru.clevertec.exception.ServiceException;
-import ru.clevertec.mapper.Mapper;
+import ru.clevertec.mapper.CardMapper;
 import ru.clevertec.repository.CardRepository;
-import ru.clevertec.task.collection.CustomArrayList;
-import ru.clevertec.task.collection.CustomList;
-import ru.clevertec.util.PropertiesUtil;
+
+import java.util.List;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class CardServiceImpl implements CardService {
 
+    @Value("${check.pageSizeDefault}")
+    private int pageSize;
+    @Value("${check.pageDefault}")
+    private int page;
     private final CardRepository repository;
-    private final Mapper<CardDto, Card> mapper;
+    private final CardMapper mapper;
 
     @Override
     @Transactional
     public CardDto save(Card card) {
-        return mapper.map(repository.save(card));
+        return mapper.toDto(repository.save(card));
     }
 
     @Override
-    public CustomList<CardDto> findAll(String pageSizeStr, String pageStr) {
-        int pageSize = PropertiesUtil.getYamlProperties().getCheck().getPageSizeDefault();
-        int page = PropertiesUtil.getYamlProperties().getCheck().getPageDefault() * pageSize;
+    public List<CardDto> findAll(String pageSizeStr, String pageStr) {
         if (pageSizeStr != null) {
             pageSize = Integer.parseInt(pageSizeStr);
         }
         if (pageStr != null) {
-            page = Integer.parseInt(pageStr) * pageSize;
+            page = Integer.parseInt(pageStr);
         }
-        CustomList<CardDto> cardDtoList = new CustomArrayList<>();
-        repository.findAll(PageRequest.of(page, pageSize)).stream()
-                .map(mapper::map)
-                .forEach(cardDtoList::add);
-        return cardDtoList;
+        return repository.findAll(PageRequest.of(page, pageSize)).stream()
+                .map(mapper::toDto)
+                .collect(toList());
     }
 
     @Override
     public CardDto findById(Integer id) {
         return repository.findById(id)
-                .map(mapper::map)
+                .map(mapper::toDto)
                 .orElseThrow(() -> new ServiceException(
                         String.format("Card with id %d not found", id)));
     }
@@ -56,7 +57,7 @@ public class CardServiceImpl implements CardService {
     @Override
     public CardDto findByNumber(String number) {
         return repository.findFirstByNumber(Integer.parseInt(number))
-                .map(mapper::map)
+                .map(mapper::toDto)
                 .orElseThrow(() -> new ServiceException(
                         String.format("Card with number %s not found", number)));
     }
@@ -64,25 +65,19 @@ public class CardServiceImpl implements CardService {
     @Override
     @Transactional
     public CardDto update(Integer id, Card card) {
-        try {
-            findById(id);
-            card.setId(id);
-            return mapper.map(repository.save(card));
-        } catch (RepositoryException e) {
-            throw new ServiceException(
-                    String.format("Card with id %d not updated", id));
+        if (repository.findById(id).isEmpty()) {
+            throw new ServiceException(String.format("Card with id %d not exist", id));
         }
+        card.setId(id);
+        return mapper.toDto(repository.save(card));
     }
 
     @Override
     @Transactional
     public void delete(Integer id) {
-        try {
-            findById(id);
-            repository.deleteById(id);
-        } catch (RepositoryException e) {
-            throw new ServiceException(
-                    String.format("Card with id %d not deleted", id));
+        if (repository.findById(id).isEmpty()) {
+            throw new ServiceException(String.format("Card with id %d not exist", id));
         }
+        repository.deleteById(id);
     }
 }
