+++ src/main/java/ru/clevertec/service/ItemServiceImpl.java
@@ -1,54 +1,55 @@
 package ru.clevertec.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.clevertec.dto.ItemDto;
 import ru.clevertec.entity.Item;
-import ru.clevertec.exception.RepositoryException;
 import ru.clevertec.exception.ServiceException;
-import ru.clevertec.mapper.Mapper;
+import ru.clevertec.mapper.ItemMapper;
 import ru.clevertec.repository.ItemRepository;
-import ru.clevertec.task.collection.CustomArrayList;
-import ru.clevertec.task.collection.CustomList;
-import ru.clevertec.util.PropertiesUtil;
+
+import java.util.List;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class ItemServiceImpl implements ItemService {
 
+    @Value("${check.pageSizeDefault}")
+    private int pageSize;
+    @Value("${check.pageDefault}")
+    private int page;
     private final ItemRepository repository;
-    private final Mapper<ItemDto, Item> mapper;
+    private final ItemMapper mapper;
 
     @Override
     @Transactional
     public ItemDto save(Item item) {
-        return mapper.map(repository.save(item));
+        return mapper.toDto(repository.save(item));
     }
 
     @Override
-    public CustomList<ItemDto> findAll(String pageSizeStr, String pageStr) {
-        int pageSize = PropertiesUtil.getYamlProperties().getCheck().getPageSizeDefault();
-        int page = PropertiesUtil.getYamlProperties().getCheck().getPageDefault() * pageSize;
+    public List<ItemDto> findAll(String pageSizeStr, String pageStr) {
         if (pageSizeStr != null) {
             pageSize = Integer.parseInt(pageSizeStr);
         }
         if (pageStr != null) {
-            page = Integer.parseInt(pageStr) * pageSize;
+            page = Integer.parseInt(pageStr);
         }
-        CustomList<ItemDto> itemDtoList = new CustomArrayList<>();
-        repository.findAll(PageRequest.of(page, pageSize)).stream()
-                .map(mapper::map)
-                .forEach(itemDtoList::add);
-        return itemDtoList;
+        return repository.findAll(PageRequest.of(page, pageSize)).stream()
+                .map(mapper::toDto)
+                .collect(toList());
     }
 
     @Override
     public ItemDto findById(Integer id) {
         return repository.findById(id)
-                .map(mapper::map)
+                .map(mapper::toDto)
                 .orElseThrow(() -> new ServiceException(
                         String.format("Item with id %d not found", id)));
     }
@@ -56,25 +57,19 @@ public class ItemServiceImpl implements ItemService {
     @Override
     @Transactional
     public ItemDto update(Integer id, Item item) {
-        try {
-            findById(id);
-            item.setId(id);
-            return mapper.map(repository.save(item));
-        } catch (RepositoryException e) {
-            throw new ServiceException(
-                    String.format("Item with id %d not updated", id));
+        if (repository.findById(id).isEmpty()) {
+            throw new ServiceException(String.format("Item with id %d not exist", id));
         }
+        item.setId(id);
+        return mapper.toDto(repository.save(item));
     }
 
     @Override
     @Transactional
     public void delete(Integer id) {
-        try {
-            findById(id);
-            repository.deleteById(id);
-        } catch (RepositoryException e) {
-            throw new ServiceException(
-                    String.format("Item with id %d not deleted", id));
+        if (repository.findById(id).isEmpty()) {
+            throw new ServiceException(String.format("Item with id %d not exist", id));
         }
+        repository.deleteById(id);
     }
 }
